# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API Documentation",
    "version": "1.0.0",
    "description": "API documentation for my Next.js app"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://story-teller-batrinutsilvius-projects.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token from the login endpoint"
      }
    }
  },
  "paths": {
...
    "/api/stories/categories/{category_id}/languages/{language_id}": {
      "get": {
        "summary": "Gets all the stories from a category by language",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The category ID"
          },
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The language ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of stories in the category with translation for the given language",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "StoryCategory ID"
                      },
                      "story": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "photo_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "storyTranslations": {
                            "type": "array",
                            "description": "Contains one translation for the requested language, sorted by title",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "integer"
                                },
                                "title": {
                                  "type": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/stories": {
      "post": {
        "summary": "Create a complete story with translation and pages",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "language_id",
                  "title",
                  "pages"
                ],
                "properties": {
                  "photo_url": {
                    "type": "string"
                  },
                  "story_series_id": {
                    "type": "integer"
                  },
                  "category_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "language_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "The Great Adventure"
                  },
                  "audio_url": {
                    "type": "string"
                  },
                  "pages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "page_number",
                        "text_content"
                      ],
                      "properties": {
                        "page_number": {
                          "type": "integer"
                        },
                        "text_content": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Story created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "photo_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "story_series_id": {
                      "type": "integer",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "storyCategories": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "category": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "categoryTranslations": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "type": "integer"
                                    },
                                    "name": {
                                      "type": "string"
                                    },
                                    "language_id": {
                                      "type": "integer"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "storyTranslations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          },
                          "language": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "name": {
                                "type": "string"
                              },
                              "country_code": {
                                "type": "string"
                              }
                            }
                          },
                          "storyPages": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "integer"
                                },
                                "page_number": {
                                  "type": "integer"
                                },
                                "text_content": {
                                  "type": "string"
                                },
                                "photo_url": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "audio_url": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields or invalid references"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin access required"
          },
          "404": {
            "description": "Language or category not found"
          }
        }
      }
    }
  },
...
}